Split generation route helpers
